Change adminController.class
