+++ web/src/test/java/ua/com/vertex/controllers/AdminControllerTest.java
@@ -2,304 +2,24 @@ package ua.com.vertex.controllers;
 
 import org.junit.Before;
 import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.mockito.Mock;
-import org.mockito.runners.MockitoJUnitRunner;
-import org.springframework.dao.DataIntegrityViolationException;
-import org.springframework.ui.ExtendedModelMap;
-import org.springframework.ui.Model;
-import org.springframework.validation.BindingResult;
 import org.springframework.web.servlet.ModelAndView;
-import ua.com.vertex.beans.*;
-import ua.com.vertex.logic.interfaces.*;
 
-import java.math.BigDecimal;
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
-
-import static org.junit.Assert.*;
-import static org.mockito.Matchers.anyInt;
-import static org.mockito.Mockito.*;
+import static org.junit.Assert.assertEquals;
 import static ua.com.vertex.controllers.AdminController.*;
-import static ua.com.vertex.controllers.CertificateDetailsPageController.ERROR;
 
-@RunWith(MockitoJUnitRunner.class)
+
 public class AdminControllerTest {
 
-    private final String MSG_INVALID_DATA = "Have wrong objects in model";
-    private final String MSG_INVALID_VIEW = "Have wrong viewName in ModelAndView";
-
     private AdminController underTest;
-    private Model model;
-    private Certificate certificate;
-    private User user;
-    private Course course;
-    private CertificateWithUserForm certificateWithUserForm;
-    private PaymentForm paymentForm;
-
-    @Mock
-    private CertificateLogic certificateLogic;
-
-    @Mock
-    private UserLogic userLogic;
-
-    @Mock
-    private CourseLogic courseLogic;
-
-    @Mock
-    private AccountingLogic accountingLogic;
-
-    @Mock
-    private PaymentLogic paymentLogic;
-
-    @Mock
-    private BindingResult bindingResult;
 
     @Before
-    public void setUp() throws Exception {
-        underTest = new AdminController(certificateLogic, userLogic, courseLogic, accountingLogic, paymentLogic);
-        model = new ExtendedModelMap();
-        certificate = new Certificate.Builder().setUserId(1).setCertificationDate(LocalDate.parse("2016-12-01"))
-                .setCourseName("Java Professional").setLanguage("Java").getInstance();
-        user = new User.Builder().setUserId(1).setEmail("test@mail.com").setFirstName("Test").setLastName("Test")
-                .getInstance();
-        course = new Course.Builder().setId(1).setName("JavaPro").
-                setStart(LocalDateTime.of(2017, 4, 25, 12, 30)).setFinished(false)
-                .setPrice(BigDecimal.valueOf(4000)).setTeacherName("Mr. Teacher").setNotes("Test").getInstance();
-        paymentForm = new PaymentForm(1, 1, new Payment.Builder().setPaymentId(1).setDealId(1)
-                .setAmount(BigDecimal.valueOf(1000))
-                .setPaymentDate(LocalDateTime.of(2017, 4, 25, 12, 30)).getInstance());
-        certificateWithUserForm = new CertificateWithUserForm();
-        certificateWithUserForm.setCertificate(certificate);
-        certificateWithUserForm.setUser(user);
+    public void setUp() {
+        underTest = new AdminController();
     }
 
     @Test
     public void adminHasCorrectDataInModel() throws Exception {
         ModelAndView result = underTest.admin();
-        assertEquals(MSG_INVALID_VIEW, result.getViewName(), ADMIN_JSP);
-    }
-
-    @Test
-    public void addCertificateWithUserIdReturnCorrectView() throws Exception {
-        ModelAndView result = underTest.addCertificateWithUserId();
-        assertEquals(MSG_INVALID_VIEW, result.getViewName(), SELECT_USER_JSP);
-    }
-
-    @Test
-    public void searchUserCalledInUserLogic() throws Exception {
-        underTest.searchUser("Test", model);
-        verify(userLogic).searchUser("Test");
-    }
-
-    @Test
-    public void searchUserHasCorrectDataInModelAndReturnCorrectViewWhenEmptyList() throws Exception {
-
-        when(userLogic.searchUser("Test")).thenReturn(new ArrayList<>());
-        assertEquals(MSG_INVALID_VIEW, underTest.searchUser("Test", model), SELECT_USER_JSP);
-        assertTrue(MSG_INVALID_DATA, model.containsAttribute(MSG));
-        assertNotNull(MSG_INVALID_DATA, model.asMap().get(MSG));
-        assertTrue(MSG_INVALID_DATA, model.containsAttribute(USERS));
-        assertNotNull(MSG_INVALID_DATA, model.containsAttribute(USERS));
-        assertTrue(MSG_INVALID_DATA, model.asMap().containsValue(LOG_USER_NOT_FOUND));
-        @SuppressWarnings("unchecked") List<User> users = (List<User>) model.asMap().get(USERS);
-        assertTrue(MSG_INVALID_DATA, users.isEmpty());
-    }
-
-    @Test
-    public void searchUserHasCorrectDataInModelAndReturnCorrectViewWhenNotEmptyList() throws Exception {
-        when(userLogic.searchUser("Test")).thenReturn(Collections.singletonList(user));
-        assertEquals(MSG_INVALID_VIEW, underTest.searchUser("Test", model), SELECT_USER_JSP);
-        verify(userLogic).searchUser("Test");
-        assertTrue(MSG_INVALID_DATA, model.containsAttribute(USERS));
-        //noinspection unchecked
-        List<User> users = (List<User>) model.asMap().get(USERS);
-        assertFalse(MSG_INVALID_DATA, users.isEmpty());
-        assertEquals(MSG_INVALID_DATA, users.get(0), user);
-    }
-
-    @Test
-    public void searchUserHasCorrectDataInModelAndReturnCorrectViewWhenEmptyInputString() throws Exception {
-        assertEquals(MSG_INVALID_VIEW, underTest.searchUser("", model), SELECT_USER_JSP);
-        assertTrue(MSG_INVALID_DATA, model.containsAttribute(MSG));
-        assertTrue(MSG_INVALID_DATA, model.asMap().containsValue(LOG_INCORRECT_DATA));
-        assertNotNull(MSG_INVALID_DATA, model.asMap().containsValue(LOG_INCORRECT_DATA));
-    }
-
-    @Test
-    public void searchUserHasCorrectDataInModelAndReturnCorrectViewWhenException() throws Exception {
-        when(userLogic.searchUser("Test")).thenThrow(new Exception("Test"));
-        assertEquals(MSG_INVALID_VIEW, underTest.searchUser("Test", model), ERROR);
-    }
-
-    @Test
-    public void selectUserHasCorrectDataInModelAndView() throws Exception {
-        ModelAndView result = underTest.selectUser(333);
-        assertEquals(MSG_INVALID_VIEW, result.getViewName(), ADD_CERTIFICATE_WITH_USER_ID_JSP);
-        assertEquals(MSG_INVALID_DATA, result.getModel().get(USER_ID), 333);
-        assertEquals(MSG_INVALID_DATA, result.getModel().get(CERTIFICATE), new Certificate());
-    }
-
-    @Test
-    public void checkCertificateWithUserIdCalledInCertificateLogic() throws Exception {
-        underTest.checkCertificateWithUserId(certificate, bindingResult, model);
-        verify(certificateLogic).addCertificate(certificate);
-    }
-
-    @Test
-    public void checkCertificateWithUserIdHasCorrectDataInModelAndReturnCorrectView() throws Exception {
-        when(certificateLogic.addCertificate(certificate)).thenReturn(333);
-        assertEquals(MSG_INVALID_VIEW, underTest.checkCertificateWithUserId(certificate, bindingResult, model)
-                , ADMIN_JSP);
-        assertTrue(MSG_INVALID_DATA, model.containsAttribute(MSG));
-        assertTrue(MSG_INVALID_DATA, model.asMap().containsValue(LOG_CERTIFICATE_ADDED + "333"));
-    }
-
-    @Test
-    public void checkCertificateWithUserIdReturnCorrectViewWhenException() throws Exception {
-        when(certificateLogic.addCertificate(certificate)).thenThrow(new Exception("Test"));
-        assertEquals(MSG_INVALID_VIEW, underTest.checkCertificateWithUserId(certificate, bindingResult, model)
-                , ERROR);
-    }
-
-    @Test
-    public void checkCertificateWithUserIdHasCorrectDataInModelAndReturnCorrectViewWhenBindigResultHasError()
-            throws Exception {
-        when(bindingResult.hasErrors()).thenReturn(true);
-        assertEquals(MSG_INVALID_VIEW, underTest.checkCertificateWithUserId(certificate, bindingResult, model)
-                , ADD_CERTIFICATE_WITH_USER_ID_JSP);
-        assertTrue(MSG_INVALID_DATA, model.containsAttribute(MSG));
-        assertTrue(MSG_INVALID_DATA, model.asMap().containsValue(LOG_INCORRECT_DATA));
-    }
-
-    @Test
-    public void addCertificateAndCreateUserReturnCorrectDataToModelAndView() throws Exception {
-        ModelAndView result = underTest.addCertificateAndCreateUser();
-        assertEquals(MSG_INVALID_VIEW, result.getViewName(), ADD_CERTIFICATE_AND_USER_JSP);
-        assertEquals(MSG_INVALID_DATA, result.getModel().get(CERTIFICATE_WITH_USER_FORM)
-                , new CertificateWithUserForm());
-    }
-
-    @Test
-    public void checkCertificateAndUserCalledInCertificateLogic() throws Exception {
-        underTest.checkCertificateAndUser(certificateWithUserForm, bindingResult, model);
-        verify(certificateLogic).addCertificateAndCreateUser(certificateWithUserForm.getCertificate()
-                , certificateWithUserForm.getUser());
-    }
-
-    @Test
-    public void checkCertificateAndUserHasCorrectDataInModelAndReturnCorrectView() throws Exception {
-        when(certificateLogic.addCertificateAndCreateUser(certificate, user)).thenReturn(333);
-
-        String returnPage = underTest.checkCertificateAndUser(certificateWithUserForm, bindingResult, model);
-
-        assertEquals(MSG_INVALID_VIEW, returnPage, ADMIN_JSP);
-        assertTrue(MSG_INVALID_DATA, model.containsAttribute(MSG));
-        assertTrue(MSG_INVALID_DATA, model.asMap().containsValue(LOG_CERTIFICATE_ADDED + "333"));
-    }
-
-    @Test
-    public void checkCertificateAndUserHasCorrectDataInModelAndReturnCorrectViewWhenDataIntegrityViolationException()
-            throws Exception {
-        when(certificateLogic.addCertificateAndCreateUser(certificate
-                , user)).thenThrow(new DataIntegrityViolationException("Test"));
-
-        assertEquals(MSG_INVALID_VIEW, underTest.checkCertificateAndUser(certificateWithUserForm, bindingResult, model)
-                , ADD_CERTIFICATE_AND_USER_JSP);
-        assertTrue(MSG_INVALID_DATA, model.containsAttribute(MSG));
-        assertTrue(MSG_INVALID_DATA, model.asMap().containsValue(LOG_INVALID_USER_EMAIL));
-    }
-
-    @Test
-    public void checkCertificateAndUserHasCorrectDataInModelAndReturnCorrectViewWhenException() throws Exception {
-        when(certificateLogic.addCertificateAndCreateUser(certificate
-                , user)).thenThrow(new Exception("Test"));
-
-        assertEquals(MSG_INVALID_VIEW, underTest.checkCertificateAndUser(certificateWithUserForm, bindingResult, model)
-                , ERROR);
-    }
-
-    @Test
-    public void checkCertificateAndUserHasCorrectDataInModelAndReturnCorrectViewWhenBindigResultHasError()
-            throws Exception {
-        when(bindingResult.hasErrors()).thenReturn(true);
-
-        assertEquals(MSG_INVALID_VIEW, underTest.checkCertificateAndUser(certificateWithUserForm, bindingResult, model)
-                , ADD_CERTIFICATE_AND_USER_JSP);
-        assertTrue(MSG_INVALID_DATA, model.containsAttribute(MSG));
-        assertTrue(MSG_INVALID_DATA, model.asMap().containsValue(LOG_INCORRECT_DATA));
-    }
-
-    @Test
-    public void selectCourseForPaymentReturnCorrectViewWhenException() throws Exception {
-        when(courseLogic.getAllCoursesWithDept()).thenThrow(new DataIntegrityViolationException("Test"));
-        assertEquals(MSG_INVALID_VIEW, underTest.selectCourseForPayment().getViewName(), ERROR);
-    }
-
-    @Test
-    public void selectCourseForPaymentReturnCorrectViewAndDataInModel() throws Exception {
-        when(courseLogic.getAllCoursesWithDept()).thenReturn(Collections.singletonList(course));
-
-        ModelAndView result = underTest.selectCourseForPayment();
-        verify(courseLogic, times(1)).getAllCoursesWithDept();
-
-        assertEquals(MSG_INVALID_VIEW, result.getViewName(), SELECT_COURSE_FOR_PAYMENT_JSP);
-        assertEquals(MSG_INVALID_DATA, result.getModel().get(COURSES), Collections.singletonList(course));
-    }
-
-    @Test
-    public void selectUserForPaymentReturnCorrectViewWhenException() throws Exception {
-        when(accountingLogic.getCourseUsers(anyInt())).thenThrow(new DataIntegrityViolationException("Test"));
-        assertEquals(MSG_INVALID_VIEW, underTest.selectUserForPayment(1).getViewName(), ERROR);
-    }
-
-    @Test
-    public void selectUserForPaymentReturnCorrectViewAndDataInModel() throws Exception {
-        when(accountingLogic.getCourseUsers(anyInt())).thenReturn(Collections.singletonList(user));
-
-        ModelAndView result = underTest.selectUserForPayment(1);
-        verify(accountingLogic, times(1)).getCourseUsers(1);
-
-        assertEquals(MSG_INVALID_VIEW, result.getViewName(), SELECT_USER_FOR_PAYMENT_JSP);
-        assertEquals(MSG_INVALID_DATA, result.getModel().get(USERS), Collections.singletonList(user));
-    }
-
-    @Test
-    public void createPaymentReturnCorrectViewWhenBindingResultHasError() throws Exception {
-        when(bindingResult.hasErrors()).thenReturn(true);
-
-        ModelAndView result = underTest.createPayment(paymentForm, bindingResult, new ModelAndView());
-
-        assertEquals(MSG_INVALID_VIEW, result.getViewName(), SELECT_USER_FOR_PAYMENT_JSP);
-
-        assertTrue(MSG_INVALID_DATA, result.getModel().containsKey(COURSE_ID_FOR_PAY));
-        assertEquals(MSG_INVALID_DATA, result.getModel().get(COURSE_ID_FOR_PAY), 1);
-
-        assertTrue(MSG_INVALID_DATA, result.getModel().containsKey(USER_ID_FOR_PAY));
-        assertEquals(MSG_INVALID_DATA, result.getModel().get(USER_ID_FOR_PAY), 1);
-    }
-
-    @Test
-    public void createPaymentReturnCorrectViewWhenException() throws Exception {
-        when(paymentLogic.createNewPaymentAndUpdateAccounting(paymentForm))
-                .thenThrow(new DataIntegrityViolationException("Test"));
-        assertEquals(MSG_INVALID_VIEW, underTest.createPayment(paymentForm, bindingResult, new ModelAndView())
-                .getViewName(), ERROR);
-    }
-
-    @Test
-    public void createPaymentReturnCorrectViewAndDataInModel() throws Exception {
-        when(paymentLogic.createNewPaymentAndUpdateAccounting(paymentForm)).thenReturn(1);
-
-        ModelAndView result = underTest.createPayment(paymentForm, bindingResult, new ModelAndView());
-        verify(paymentLogic, times(1)).createNewPaymentAndUpdateAccounting(paymentForm);
-
-        assertEquals(MSG_INVALID_VIEW, result.getViewName(), ADMIN_JSP);
-        assertTrue(MSG_INVALID_DATA, result.getModel().containsKey(MSG));
-        assertEquals(MSG, result.getModel().get(MSG), "Payment create successful!!!");
-
+        assertEquals("Have wrong viewName in ModelAndView", result.getViewName(), ADMIN_JSP);
     }
 }
